Maps/Warnings: reduce the amount of warnings/err...

+++ src/tools/ToolsCata/vmap_tools/vmap4_extractor/mpqfile.cpp
@@ -58,7 +58,7 @@ MPQFile::MPQFile(HANDLE mpq, const char* filename, bool warnNoExist /*= true*/) 
 
     DWORD read = 0;
     buffer = new char[size];
-    if (!SFileReadFile(file, buffer, size, &read) || size != read)
+    if (!SFileReadFile(file, buffer, (DWORD)size, &read) || size != read)
     {
         fprintf(stderr, "Can't read %s, size=%u read=%u!\n", filename, uint32(size), uint32(read));
         SFileCloseFile(file);
